+++ R/agentset-functions.R
@@ -676,7 +676,7 @@ setMethod(
   signature = c("matrix", "worldMatrix", "missing"),
   definition = function(agents, world) {
     maxAgents <- withMax(world = world, agents = agents)
-    row <- sample(1:NROW(maxAgents), size = 1)
+    row <- resample(1:NROW(maxAgents), size = 1)
     return(maxAgents[row, , drop = FALSE])
   }
 )
@@ -688,7 +688,7 @@ setMethod(
   signature = c("matrix", "worldArray", "character"),
   definition = function(agents, world, var) {
     maxAgents <- withMax(world = world, agents = agents, var = var)
-    row <- sample(1:NROW(maxAgents), size = 1)
+    row <- resample(1:NROW(maxAgents), size = 1)
     return(maxAgents[row, , drop = FALSE])
   }
 )
@@ -700,7 +700,7 @@ setMethod(
   signature = c("agentMatrix", "missing", "character"),
   definition = function(agents, var) {
     maxAgents <- withMax(agents = agents, var = var)
-    row <- sample(1:NLcount(maxAgents), size = 1)
+    row <- resample(1:NLcount(maxAgents), size = 1)
     return(maxAgents[row, ])
   }
 )
@@ -770,7 +770,7 @@ setMethod(
   signature = c("matrix", "worldMatrix", "missing"),
   definition = function(agents, world) {
     minAgents <- withMin(world = world, agents = agents)
-    row <- sample(1:NROW(minAgents), size = 1)
+    row <- resample(1:NROW(minAgents), size = 1)
     return(minAgents[row, , drop = FALSE])
   }
 )
@@ -782,7 +782,7 @@ setMethod(
   signature = c("matrix", "worldArray", "character"),
   definition = function(agents, world, var) {
     minAgents <- withMin(world = world, agents = agents, var = var)
-    row <- sample(1:NROW(minAgents), size = 1)
+    row <- resample(1:NROW(minAgents), size = 1)
     return(minAgents[row, , drop = FALSE])
   }
 )
@@ -794,7 +794,7 @@ setMethod(
   signature = c("agentMatrix", "missing", "character"),
   definition = function(agents, var) {
     minAgents <- withMin(agents = agents, var = var)
-    row <- sample(1:NLcount(minAgents), size = 1)
+    row <- resample(1:NLcount(minAgents), size = 1)
     return(minAgents[row, ])
   }
 )
@@ -982,14 +982,14 @@ setMethod(
 
 
     if (inherits(agents, "agentMatrix")) {
-      row <- sample(1:NROW(agents), size = n, replace = FALSE)
+      row <- resample(1:NROW(agents), size = n, replace = FALSE)
       row <- row[order(row)]
       turtles <- agents[row, ]
       return(turtles)
     } else {
       if (ncol(agents) == 2 & colnames(agents)[1] == "pxcor") {
         # patches
-        row <- sample(1:NROW(agents), size = n, replace = FALSE)
+        row <- resample(1:NROW(agents), size = n, replace = FALSE)
         row <- row[order(row)]
         patches <- agents[row, , drop = FALSE]
         return(patches)
@@ -1001,13 +1001,13 @@ setMethod(
           if (ncol(agents) == 3) {
             # patches with id
             row <- tapply(X = 1:nrow(agents), INDEX = as.factor(agents[, "id"]),
-                          FUN = function(x) sample(x, size = n, replace = FALSE))
+                          FUN = function(x) resample(x, size = n, replace = FALSE))
             patches <- agents[unlist(row), c("pxcor", "pycor")]
             return(patches)
           } else {
             # whoNUmbers of turtles with id
             row <- tapply(X = 1:nrow(agents), INDEX = as.factor(agents[, "id"]),
-                          FUN = function(x) sample(x, size = n, replace = FALSE))
+                          FUN = function(x) resample(x, size = n, replace = FALSE))
             turtles <- agents[unlist(row), "whoTurtles"]
             return(turtles)
           }
@@ -1072,7 +1072,6 @@ setMethod(
 #'
 #'
 #' @export
-#' @importFrom Hmisc mApply
 #' @rdname oneOf
 #'
 #' @author Sarah Bauduin
@@ -1089,23 +1088,28 @@ setMethod(
   "oneOf",
   signature = c("matrix"),
   definition = function(agents) {
-    if (inherits(agents, "agentMatrix")) {
-      nOf(agents = agents, n = 1)
-    } else {
-      if (ncol(agents) == 2 & colnames(agents)[1] == "pxcor") {
-        # patches
-        nOf(agents = agents, n = 1)
-      } else if (ncol(agents) == 3) {
-        # patches with id
-        mApply(X = agents[, c("pxcor", "pycor")], INDEX = as.factor(agents[, "id"]),
-               FUN = oneOf, keepmatrix = TRUE)
-      } else {
-        # whoNUmbers of turtles with id
-        whoTurtles <- tapply(X = agents[, "whoTurtles"], INDEX = as.factor(agents[, "id"]),
-                             FUN = function(x) ifelse(length(x) == 1, x, sample(x, size = 1)))
-        return(as.numeric(whoTurtles))
-      }
-    }
+    nOf(agents = agents, n = 1)
+    # if (inherits(agents, "agentMatrix")) {
+    #   nOf(agents = agents, n = 1)
+    # } else {
+    #   nO
+    #   if (ncol(agents) == 2 & colnames(agents)[1] == "pxcor") {
+    #     # patches
+    #     nOf(agents = agents, n = 1)
+    #   } else if (ncol(agents) == 3) {
+    #     # patches with id
+    #     out <- nOf(agents = agents, n = 1)
+    #     # out <- mApply(X = agents[, c("pxcor", "pycor")], INDEX = as.factor(agents[, "id"]),
+    #     #        FUN = oneOf, keepmatrix = TRUE)
+    #     return(out)
+    #
+    #   } else {
+    #     # whoNUmbers of turtles with id
+    #     whoTurtles <- tapply(X = agents[, "whoTurtles"], INDEX = as.factor(agents[, "id"]),
+    #                          FUN = function(x) resample(x, size = 1) )
+    #     return(as.numeric(whoTurtles))
+    #   }
+    # }
   }
 )
 
@@ -1809,9 +1813,14 @@ setMethod(
           # levels in the order of the unique old levels numbers
           turtlesLevelsVarUpdated <- turtlesLevelsVar[unique(turtlesVar)[order(turtlesVarUnique)]]
           turtles@levels[[var]] <- turtlesLevelsVarUpdated
+
           # replace the levels number starting to 1 and increasing by 1
-          turtlesVarUpdated <- mapvalues(x = turtlesVar, from = turtlesVarUnique,
+          # turtlesVarUpdated <- mapvalues(x = turtlesVar, from = turtlesVarUnique,
+          #                                to = rank(turtlesVarUnique))
+          turtlesVarUpdated <- rename(x = turtlesVar, from = turtlesVarUnique,
                                          to = rank(turtlesVarUnique))
+          # if (!identical(turtlesVarUpdated, a)) stop("mapvalues and rename different in NLset")
+
           turtles@.Data[, var] <- turtlesVarUpdated
 
           if (identical(agents, turtles)) {
@@ -1847,8 +1856,13 @@ setMethod(
                                                         [order(turtlesVarUnique)]]
             turtles@levels[[var[varLevels]]] <- turtlesLevelsVarUpdated
             # replace the levels number starting to 1 and increasing by 1
-            turtlesVarUpdated <- mapvalues(x = turtlesVar, from = turtlesVarUnique,
-                                           to = rank(turtlesVarUnique))
+            # turtlesVarUpdated <- mapvalues(x = turtlesVar, from = turtlesVarUnique,
+            #                                to = rank(turtlesVarUnique))
+            turtlesVarUpdated <- rename(x = turtlesVar, from = turtlesVarUnique,
+                        to = rank(turtlesVarUnique))
+            # if (!identical(turtlesVarUpdated, a)) stop("mapvalues and rename different in NLset 2")
+
+
             turtles@.Data[, var[varLevels]] <- turtlesVarUpdated
 
             if (identical(agents, turtles)) {
@@ -1872,8 +1886,14 @@ setMethod(
                                                             [order(turtlesVarUnique)]]
                 turtles@levels[[var[i]]] <- turtlesLevelsVarUpdated
                 # replace the levels number starting to 1 and increasing by 1
-                turtlesVarUpdated <- mapvalues(x = turtlesVar, from = turtlesVarUnique,
-                                               to = rank(turtlesVarUnique))
+                # browser()
+                # turtlesVarUpdated <- mapvalues(x = turtlesVar, from = turtlesVarUnique,
+                #                                to = rank(turtlesVarUnique))
+                turtlesVarUpdated <- rename(x = turtlesVar, from = turtlesVarUnique,
+                            to = rank(turtlesVarUnique))
+                # if (!identical(turtlesVarUpdated, a)) stop("mapvalues and rename different in NLset")
+
+
                 turtles@.Data[, var[i]] <- turtlesVarUpdated
 
                 turtles[, var[i]] <- as.character(val[, var[i]])
@@ -1891,8 +1911,14 @@ setMethod(
                                                             [order(turtlesVarUnique)]]
                 turtles@levels[[var[i]]] <- turtlesLevelsVarUpdated
                 # replace the levels number starting to 1 and increasing by 1
-                turtlesVarUpdated <- mapvalues(x = turtlesVar, from = turtlesVarUnique,
-                                               to = rank(turtlesVarUnique))
+                # turtlesVarUpdated <- mapvalues(x = turtlesVar, from = turtlesVarUnique,
+                #                                to = rank(turtlesVarUnique))
+                turtlesVarUpdated <- rename(x = turtlesVar, from = turtlesVarUnique,
+                            to = rank(turtlesVarUnique))
+                # if (!identical(turtlesVarUpdated, a)) stop("mapvalues and rename different in NLset")
+
+
+
                 turtles@.Data[, var[i]] <- turtlesVarUpdated
 
                 turtles[iAgents, var[i]] <- as.character(val[, i])