+++ NEWS.md
@@ -1,5 +1,17 @@
 Known issues: <https://github.com/PredictiveEcology/NetLogoR/issues>
 
+Version 0.3.10
+=============
+
+# Dependency changes
+* removed support for R 3.6. To our knowledge, `NetLogoR` will still work with R 3.6, but we no longer support it.
+* removed dependencies: `car`, `Hmisc`, `plyr`
+
+# Bugfixes
+* using sample in several agent functions would have given the wrong sample (see `?sample`) when the length of the argument `x` is 1.
+  This occurred under some (but now all) conditions where there are multiple agents with the same `id`, along with some agents with only one member within an `id`, and where the objective is to select one of the using `oneOf` or `nOf(agent, n = 1)`. Now we use `resample` as defined in the examples of `?base::sample`.
+* minor bugfixes
+
 Version 0.3.9
 =============